work around a flaky test
Without the little pause at the beginning, the workflow menu may apparently
not be loaded completely before opened, and not load completely after that
either. If that happens (maybe half of the time in my dev env), the submit
button will not be found and the test raises an error.